ndb: add threaded ingester and writer queues
This adds an ingester and writer queue to nostrdb.

Ingester: used for queueing json for processing
Writer: A single-thread for writing parsed notes to the database.

For now there is only one ingester thread, but in the future was can
expand this into a thread pool for max CPU utilization.

This also implements a share-nothing architecture similar to strfry.
These queues simply pass around owned pointers, thread contention is
minimized since locks held on queues are very short lived.