stats: add reaction counter
Implement the initial nostrdb stats counter. We start by creating a
simple reaction counter that increases the reaction count in the note
metadata whenever it sees a new like for a note. It counts any like,
even - atm which is probably wrong. In the future we can fix these with
migrations.

There is a race condition if there are duplicate notes in the same
ingestor block, so we should fix that some time in the future. I added a
note on the ndb_write_reaction_stats function for future bug fixers.

This will be used to replace the EventCounter logic in Damus iOS. No
longer will we need to maintain an in-memory counter which is always out
of date. Now we will have proper stats when scrolling through the feed,
at least after the likes have been fetched...

Changelog-Added: Add reaction counts to note metadata