Clean up direct access to rcur.
A few places where we reached into rcur directly can now be removed.

This means we save the pointer to the initial content, and have
helpers to transfer into ndb_str_block.

URL parsing is a bit awkward, but it always was.

parse_nostr_bech32 (only used in test code) is altered too.

Signed-off-by: Rusty Russell <rusty@rustcorp.com.au>